Tried to calculate thread time

+++ SharedComponents/Module/IModuleInfo.cs
@@ -2,6 +2,6 @@ namespace SharedComponents.Module
 {
     public interface IModuleInfo
     {
-        int Cpu { get; }
+        double Cpu { get; }
     }
 }
