+++ SharedComponents/Module/ModuleInfo.cs
@@ -2,6 +2,6 @@ namespace SharedComponents.Module
 {
     public class ModuleInfo : IModuleInfo
     {
-        public int Cpu { get; set; }
+        public double Cpu { get; set; }
     }
 }
